+++ tests/integration_tests/test_sqlalchemy/test_reflect.py
@@ -3,7 +3,7 @@ import sqlalchemy as db
 from sqlalchemy.engine import Engine
 
 from clickhouse_connect import common
-from clickhouse_connect.cc_sqlalchemy.datatypes.sqltypes import UInt32, SimpleAggregateFunction
+from clickhouse_connect.cc_sqlalchemy.datatypes.sqltypes import UInt32, SimpleAggregateFunction, Point
 
 
 def test_basic_reflection(test_engine: Engine):
@@ -29,6 +29,20 @@ def test_full_table_reflection(test_engine: Engine, test_db: str):
     table = db.Table('reflect_test', metadata, autoload_with=test_engine)
     assert table.columns.key.type.__class__ == UInt32
     assert table.columns.agg.type.__class__ == SimpleAggregateFunction
+    assert 'MergeTree' in table.engine.name
+
+
+def test_types_reflection(test_engine: Engine, test_db: str):
+    common.set_setting('invalid_setting_action', 'drop')
+    conn = test_engine.connect()
+    conn.execute(f'DROP TABLE IF EXISTS {test_db}.sqlalchemy_types_test')
+    conn.execute(
+        f'CREATE TABLE {test_db}.sqlalchemy_types_test (key UInt32, pt Point) ' +
+        'ENGINE MergeTree ORDER BY key')
+    metadata = db.MetaData(bind=test_engine, schema=test_db)
+    table = db.Table('sqlalchemy_types_test', metadata, autoload_with=test_engine)
+    assert table.columns.key.type.__class__ == UInt32
+    assert table.columns.pt.type.__class__ == Point
     assert 'MergeTree' in table.engine.name