+++ CHANGELOG.md
@@ -11,6 +11,8 @@ v0.5.25, which dynamically loads the EngineSpec from the clickhouse-connect proj
 
 ## 0.7.5, 2024-03-28
 ### Bug Fixes
+- Fixed client side binding for Python format strings using `%d` (int) and `%f` (float) format patterns.  Closes
+https://github.com/ClickHouse/clickhouse-connect/issues/327
 - Allows empty `data` argument in the initializer of `ExternalFile` / `ExternalData` objects. Thanks to
   [martijnthe](https://github.com/martijnthe) for the PR!